refactor: use class and rules to handle checks ; remove duplications and error
My refactoring suggesting.
I introduced classes for the Settings and Rules that now handle checks (for example evaluation of viewMode etc.), Validation and Data changes (attributes, setting merges, pattern changes across rules).
As such, I have reduced code duplication and removed errors where for example the pattern was hardcoded.
This also makes it easier to spot where what happens. What is done and why it's done is also mixed less.

The classes have in some cases slightly different attribute names to make obvious what they exactly are (the string of the regex is titled as such, the getter of regex returns the parsed regex for the rule).
The saved settings file stays the same format (.serialize()).
No need to check for undefined for some attributes as the class now handles these empty consequently.